Backend/requirements.txt
Flask
gdown
numpy
Pillow
tensorflow
gunicorn
python-dotenv
sklearn
tqdm